+++ src/scipp/experimental/plotting/limits.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
-from .. import scalar, Variable
-from .. import abs as abs_
+from ... import scalar, Variable
+from ... import abs as abs_
 
 import numpy as np
 from typing import Tuple
