+++ src/scipp/experimental/plotting/main.py
