+++ src/scipp/experimental/plotting/mesh.py
@@ -1,11 +1,11 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
-from .. import config
-from .. import broadcast, DataArray
+from ... import config
+from ... import broadcast, DataArray
 from .limits import find_limits, fix_empty_range
 from .tools import get_cmap
-from ..utils import name_with_unit
+from ...utils import name_with_unit
 
 from functools import reduce
 from matplotlib.colors import Normalize, LogNorm
