+++ src/scipp/experimental/plotting/plot.py
