+++ src/scipp/experimental/plotting/toolbar.py
