+++ src/scipp/experimental/plotting/tools.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
-from .. import config
+from ... import config
 
 from copy import copy
 from matplotlib.colors import LinearSegmentedColormap
