+++ src/scipp/experimental/plotting/view.py
