+++ src/scipp/experimental/plotting/widgets/__init__.py
