+++ src/scipp/experimental/plotting/widgets/checkboxes.py
