+++ src/scipp/experimental/plotting/widgets/slice.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
-from ... import DataArray
-from ...utils import value_to_string
+from .... import DataArray
+from ....utils import value_to_string
 from ..view import View
 from ..model import node
