+++ src/scipp/experimental/plotting/widgets/widgetnode.py
