+++ tests/experimental/plotting/plot_test.py
@@ -2,9 +2,9 @@
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
 import scipp as sc
-from scipp.plt import Plot, Figure, widgets, input_node, node, plot
-from scipp.plt.widgets import widget_node
-from ..factory import make_dense_data_array, make_dense_dataset
+from scipp.experimental.plotting import Plot, Figure, widgets, input_node, node, plot
+from scipp.experimental.plotting.widgets import widget_node
+from ...factory import make_dense_data_array, make_dense_dataset
 import matplotlib
 import ipywidgets as ipw
 
@@ -161,7 +161,9 @@ def test_plot_3d_image_slicer_with_connected_side_histograms():
 
 def test_plot_wrapper():
     plot(make_dense_data_array(ndim=1))
-    plot(make_dense_data_array(ndim=2))
+    da = make_dense_data_array(ndim=2)
+    plot(da)
+    plot(da.data)
     ds = make_dense_dataset(ndim=1)
     plot(ds)
     plot({'a': ds['a'], 'b': ds['b']})
